fix(public-cloud.storages): error messages and code-review comments
 Apply code-review comments. Addition of error messages for Name field validation. Bug fix to display and disable Name field while editing Pool, fix to download CA Certificate.
 ref: MANAGER-8026
