+++ packages/manager/modules/pci/src/projects/project/storages/databases/database.service.js
@@ -815,28 +815,34 @@ export default class DatabaseService {
       .then(({ data }) => data);
   }
 
-  createConnectionPool(projectId, databaseId, payload) {
+  createConnectionPool(projectId, engine, databaseId, payload) {
     return this.$http
       .post(
-        `/cloud/project/${projectId}/database/postgresql/${databaseId}/connectionPool`,
+        `/cloud/project/${projectId}/database/${engine}/${databaseId}/connectionPool`,
         payload,
       )
       .then(({ data }) => data);
   }
 
-  updateConnectionPool(projectId, databaseId, connectionPoolId, payload) {
+  updateConnectionPool(
+    projectId,
+    engine,
+    databaseId,
+    connectionPoolId,
+    payload,
+  ) {
     return this.$http
       .put(
-        `/cloud/project/${projectId}/database/postgresql/${databaseId}/connectionPool/${connectionPoolId}`,
+        `/cloud/project/${projectId}/database/${engine}/${databaseId}/connectionPool/${connectionPoolId}`,
         payload,
       )
       .then(({ data }) => data);
   }
 
-  terminateConnectionPool(projectId, databaseId, connectionPoolId) {
+  terminateConnectionPool(projectId, engine, databaseId, connectionPoolId) {
     return this.$http
       .delete(
-        `/cloud/project/${projectId}/database/postgresql/${databaseId}/connectionPool/${connectionPoolId}`,
+        `/cloud/project/${projectId}/database/${engine}/${databaseId}/connectionPool/${connectionPoolId}`,
       )
       .then(({ data }) => data);
   }
