--- packages/manager/modules/pci/src/projects/project/storages/databases/database/pools/add-edit/index.js
@@ -1,36 +1,0 @@
-import angular from 'angular';
-import '@uirouter/angularjs';
-import 'oclazyload';
-
-const moduleName = 'ovhManagerPciStoragesDatabasePoolsAddEditLazyLoading';
-
-angular.module(moduleName, ['oc.lazyLoad', 'ui.router']).config(
-  /* @ngInject */ ($stateProvider) => {
-    $stateProvider
-      .state('pci.projects.project.storages.databases.dashboard.pools.add.**', {
-        url: '/add',
-        lazyLoad: ($transition$) => {
-          const $ocLazyLoad = $transition$.injector().get('$ocLazyLoad');
-
-          return import('./add-edit.module').then((mod) =>
-            $ocLazyLoad.inject(mod.default || mod),
-          );
-        },
-      })
-      .state(
-        'pci.projects.project.storages.databases.dashboard.pools.edit.**',
-        {
-          url: '/:poolId/edit',
-          lazyLoad: ($transition$) => {
-            const $ocLazyLoad = $transition$.injector().get('$ocLazyLoad');
-
-            return import('./add-edit.module').then((mod) =>
-              $ocLazyLoad.inject(mod.default || mod),
-            );
-          },
-        },
-      );
-  },
-);
-
-export default moduleName;
