+++ packages/manager/modules/pci/src/projects/project/storages/databases/database/pools/add/translations/Messages_fr_FR.json
@@ -1,6 +1,5 @@
 {
   "pci_databases_pools_add_pool_title": "Ajouter un Pool",
-  "pci_databases_pools_modify_pool_title": "Modifier un Pool",
   "pci_databases_pools_add_pool_name": "Nom",
   "pci_databases_pools_add_database": "Base de données",
   "pci_databases_pools_add_pool_mode": "Mode",
@@ -10,6 +9,6 @@
   "pci_databases_pools_add_pool_confirm": "Confirmer",
   "pci_databases_pools_add_pool_success": "Votre Pool {{name}} a été créé avec succès.",
   "pci_databases_pools_add_pool_error": "Une erreur s'est produite lors de la création de votre pool : {{message}}",
-  "pci_databases_pools_modify_pool_success": "Votre Pool {{name}} a bien été modifié.",
-  "pci_databases_pools_modify_pool_error": "Un erreur est survenue lors de la modification de votre pool {{name}} : {{message}}"
+  "pci_databases_pools_add_pool_name_required_error": "Veuillez compléter ce champ.",
+  "pci_databases_pools_add_pool_name_pattern_error": "Veuillez saisir un format valide. Doit uniquement contenir des nombres, lettres, underscores ou tirets, ne doit pas commencer par un tiret et ne doit pas être vide. Max 60 caractères."
 }
