+++ packages/manager/modules/pci/src/projects/project/storages/databases/database/pools/edit/edit.module.js
@@ -1,21 +1,23 @@
 import angular from 'angular';
+import '@uirouter/angularjs';
 import 'angular-translate';
 import '@ovh-ux/ng-translate-async-loader';
 import '@ovh-ux/ui-kit';
 
-import component from './terminate.component';
-import routing from './terminate.routing';
+import editComponent from './edit.component';
+import routing from './edit.routing';
 
-const moduleName = 'ovhManagerPciStoragesDatabasePoolsTerminate';
+const moduleName = 'ovhManagerPciStoragesDatabasePoolsEdit';
 
 angular
   .module(moduleName, [
     'ngTranslateAsyncLoader',
     'pascalprecht.translate',
     'oui',
+    'ui.router',
   ])
   .config(routing)
-  .component('ovhManagerPciStoragesDatabasePoolsAddEditComponent', component)
+  .component('ovhManagerPciStoragesDatabasePoolsEditComponent', editComponent)
   .run(/* @ngTranslationsInject:json ./translations */);
 
 export default moduleName;
