+++ packages/manager/modules/pci/src/projects/project/storages/databases/database/pools/edit/translations/Messages_fr_FR.json
@@ -1,0 +1,14 @@
+{
+  "pci_databases_pools_modify_pool_title": "Modifier un Pool",
+  "pci_databases_pools_modify_pool_name": "Nom",
+  "pci_databases_pools_modify_database": "Base de données",
+  "pci_databases_pools_modify_pool_mode": "Mode",
+  "pci_databases_pools_modify_pool_size": "Taille",
+  "pci_databases_pools_modify_pool_user": "Nom d'utilisateur",
+  "pci_databases_pools_modify_pool_cancel": "Annuler",
+  "pci_databases_pools_modify_pool_confirm": "Confirmer",
+  "pci_databases_pools_modify_pool_success": "Votre Pool {{name}} a bien été modifié.",
+  "pci_databases_pools_modify_pool_error": "Une erreur est survenue lors de la modification de votre pool {{name}} : {{message}}",
+  "pci_databases_pools_modify_pool_name_required_error": "Veuillez compléter ce champ.",
+  "pci_databases_pools_modify_pool_name_pattern_error": "Veuillez saisir un format valide. Doit uniquement contenir des nombres, lettres, underscores ou tirets, ne doit pas commencer par un tiret et ne doit pas être vide. Max 60 caractères."
+}
