--- packages/manager/modules/pci/src/projects/project/storages/databases/database/pools/information/index.scss
@@ -1,8 +1,0 @@
-.information {
-  .copy-clipboard {
-    display: flex;
-    flex: 1;
-    flex-wrap: wrap;
-    column-gap: 19rem;
-  }
-}
