+++ packages/manager/modules/pci/src/projects/project/storages/databases/database/pools/information/information.controller.js
@@ -1,16 +1,11 @@
 import { CERTIFICATE_FILENAME } from '../../../databases.constants';
 
-export default class {
+export default class PoolsInformationController {
   /* @ngInject */
   constructor(CucControllerHelper) {
     this.CucControllerHelper = CucControllerHelper;
   }
 
-  /**
-   *
-   * @param {*} database
-   * @returns downloads file ca.pem
-   */
   downloadCertificate(database) {
     return this.CucControllerHelper.constructor.downloadContent({
       fileContent: database.certificate?.ca,
