+++ packages/manager/modules/pci/src/projects/project/storages/databases/database/pools/information/translations/Messages_fr_FR.json
@@ -1,10 +1,11 @@
 {
   "pci_databases_pools_information_title": "Informations du Pool",
-  "pci_databases_pools_information_database": "Base de donnée",
+  "pci_databases_pools_information_database": "Base de données",
   "pci_databases_pools_information_port": "Port",
   "pci_databases_pools_information_ssl_mode": "Mode SSL",
   "pci_databases_pools_information_certificate_ca": "Certificat CA",
   "pci_databases_pools_information_service_uri": "Service URI",
   "pci_databases_pools_information_pool_mode": "Mode",
-  "pci_databases_pools_information_pool_size": "Taille"
+  "pci_databases_pools_information_pool_size": "Taille",
+  "pci_databases_pools_information_certificate_ca_download": "Télécharger le Certificat CA"
 }
