+++ packages/manager/modules/pci/src/projects/project/storages/databases/database/pools/pools.module.js
@@ -6,7 +6,8 @@ import '@ovh-ux/ui-kit';
 
 import poolsComponent from './pools.component';
 import routing from './pools.routing';
-import addEdit from './add-edit';
+import add from './add';
+import edit from './edit';
 import information from './information';
 import terminate from './terminate';
 
@@ -18,7 +19,8 @@ angular
     'pascalprecht.translate',
     'oui',
     'ui.router',
-    addEdit,
+    add,
+    edit,
     information,
     terminate,
   ])
